# api/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: Workflow orchestration platform for financial validation
  termsOfService: http://swagger.io/terms/
  title: Flowker API
  version: 1.0.0
servers:
- url: //localhost:4021/
paths:
...
  /v1/catalog/providers/{id}:
    get:
      description: Returns provider metadata and its JSON Schema for configuration
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_adapters_http_in_catalog.ProviderDetail'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/commons.Response'
          description: Not Found
      summary: Get provider
      tags:
      - Catalog
...
  /v1/provider-configurations:
    get:
      description: Retrieves a paginated list of provider configurations with optional
        filtering
      parameters:
      - description: Filter by status
        in: query
        name: status
        schema:
          enum:
          - active
          - disabled
          type: string
      - description: Filter by provider ID
        in: query
        name: providerId
        schema:
          type: string
      - description: Number of items per page
        in: query
        name: limit
        schema:
          default: 10
          maximum: 100
          minimum: 1
          type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        schema:
          type: string
      - description: Sort field
        in: query
        name: sortBy
        schema:
          default: createdAt
          enum:
          - createdAt
          - updatedAt
          - name
          type: string
      - description: Sort order
        in: query
        name: sortOrder
        schema:
          default: DESC
          enum:
          - ASC
          - DESC
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.ProviderConfigurationListOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: List provider configurations
      tags:
      - Provider Configurations
    post:
      description: Creates a new provider configuration validated against the provider's
        JSON Schema
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.CreateProviderConfigurationInput'
        description: Provider Configuration Input
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.ProviderConfigurationCreateOutput'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Create a provider configuration
      tags:
      - Provider Configurations
      x-codegen-request-body-name: providerConfig
  /v1/provider-configurations/{id}:
    delete:
      description: Deletes a provider configuration
      parameters:
      - description: Provider Configuration ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "204":
          content: {}
          description: No Content
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Not Found
        "409":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Conflict
        "422":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unprocessable Entity
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Delete a provider configuration
      tags:
      - Provider Configurations
    get:
      description: Retrieves a provider configuration by its ID
      parameters:
      - description: Provider Configuration ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.ProviderConfigurationOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Get provider configuration by ID
      tags:
      - Provider Configurations
    patch:
      description: Updates an existing provider configuration. If config is changed,
        re-validates against provider JSON Schema.
      parameters:
      - description: Provider Configuration ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.UpdateProviderConfigurationInput'
        description: Updated provider configuration
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.ProviderConfigurationOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Update a provider configuration
      tags:
      - Provider Configurations
      x-codegen-request-body-name: providerConfig
  /v1/provider-configurations/{id}/disable:
    post:
      description: Transitions a provider configuration from active to disabled status
      parameters:
      - description: Provider Configuration ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.ProviderConfigurationOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Disable a provider configuration
      tags:
      - Provider Configurations
  /v1/provider-configurations/{id}/enable:
    post:
      description: Transitions a provider configuration from disabled to active status
      parameters:
      - description: Provider Configuration ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.ProviderConfigurationOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Enable a provider configuration
      tags:
      - Provider Configurations
  /v1/provider-configurations/{id}/test:
    post:
      description: Tests connectivity, authentication, and end-to-end communication
        with the provider
      parameters:
      - description: Provider Configuration ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.ProviderConfigTestResultOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Not Found
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Test provider configuration connectivity
      tags:
      - Provider Configurations
...
components:
  schemas:
    api.ErrorResponse:
      properties:
        code:
          example: FLK-0001
          type: string
        message:
          example: Invalid input provided
          type: string
        title:
          example: Bad Request
          type: string
      required:
      - code
      - message
      - title
      type: object
    commons.Response:
      properties:
        code:
          type: string
        entityType:
          type: string
        err:
          type: object
        message:
          type: string
        title:
          type: string
      type: object
...
    github_com_LerianStudio_flowker_pkg_model.CreateProviderConfigurationInput:
      properties:
        config:
          additionalProperties:
            type: object
          type: object
        description:
          maxLength: 500
          type: string
        metadata:
          additionalProperties:
            type: object
          type: object
        name:
          maxLength: 100
          minLength: 1
          type: string
        providerId:
          type: string
      required:
      - config
      - name
      - providerId
      type: object
...
    github_com_LerianStudio_flowker_pkg_model.ProviderConfigurationCreateOutput:
      example:
        createdAt: createdAt
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        status: status
      properties:
        createdAt:
          type: string
        id:
          format: uuid
          type: string
        name:
          type: string
        status:
          type: string
      type: object
    github_com_LerianStudio_flowker_pkg_model.ProviderConfigurationListOutput:
      example:
        nextCursor: nextCursor
        hasMore: true
        items:
        - createdAt: createdAt
          metadata:
            key: '{}'
          providerId: providerId
          name: name
          description: description
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          config:
            key: '{}'
          status: status
          updatedAt: updatedAt
        - createdAt: createdAt
          metadata:
            key: '{}'
          providerId: providerId
          name: name
          description: description
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          config:
            key: '{}'
          status: status
          updatedAt: updatedAt
      properties:
        hasMore:
          type: boolean
        items:
          items:
            $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.ProviderConfigurationOutput'
          type: array
        nextCursor:
          type: string
      type: object
    github_com_LerianStudio_flowker_pkg_model.ProviderConfigurationOutput:
      example:
        createdAt: createdAt
        metadata:
          key: '{}'
        providerId: providerId
        name: name
        description: description
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        config:
          key: '{}'
        status: status
        updatedAt: updatedAt
      properties:
        config:
          additionalProperties:
            type: object
          type: object
        createdAt:
          type: string
        description:
          type: string
        id:
          format: uuid
          type: string
        metadata:
          additionalProperties:
            type: object
          type: object
        name:
          type: string
        providerId:
          type: string
        status:
          type: string
        updatedAt:
          type: string
      type: object
...
    github_com_LerianStudio_flowker_pkg_model.UpdateProviderConfigurationInput:
      properties:
        config:
          additionalProperties:
            type: object
          type: object
        description:
          maxLength: 500
          type: string
        metadata:
          additionalProperties:
            type: object
          type: object
        name:
          maxLength: 100
          minLength: 1
          type: string
      type: object
...
    internal_adapters_http_in_catalog.ProviderDetail:
      example:
        configSchema: configSchema
        name: name
        description: description
        id: id
        version: version
      properties:
        configSchema:
          type: string
        description:
          type: string
        id:
          type: string
        name:
          type: string
        version:
          type: string
      type: object
...